refactor: switch to a argparse + yaml config approach
The current approach of using env vars was somewhat cumbersome
so the scripts now use argparse instead.

Also the assign scripts take a somewhat easier approach:

They find which sprint does not have two or more issues inside
and start creating issues from there on.
If some people cannot be assigned this is currently printed
to the command line.

This removes the necessity to give a start sprint as all current
and future sprints are searched automatically.

It also makes the whole checking easier.